+++ tests/testthat/test-defer.R
@@ -1,10 +1,12 @@
 test_that("errors", {
+  skip_if_no_srcrefs()
   fun <- function() {
     defer(1 + "")
   }
 
   expect_snapshot(
     error = TRUE,
-    fun()
+    fun(),
+    transform = transform_column_number
   )
 })