+++ tests/testthat/test-subprocess.R
@@ -1,7 +1,9 @@
 test_that("events are properly generated", {
   skip_on_cran()
   ## This needs callr >= 3.0.0.90001, which is not yet on CRAN
-  if (packageVersion("callr") < "3.0.0.9001") skip("Need newer callr")
+  if (packageVersion("callr") < "3.0.0.9001") {
+    skip("Need newer callr")
+  }
 
   do <- function() {
     options(cli.message_class = "callr_message")
@@ -42,7 +44,9 @@ test_that("events are properly generated", {
 test_that("subprocess with default handler", {
   skip_on_cran()
   ## This needs callr >= 3.0.0.90001, which is not yet on CRAN
-  if (packageVersion("callr") < "3.0.0.9001") skip("Need newer callr")
+  if (packageVersion("callr") < "3.0.0.9001") {
+    skip("Need newer callr")
+  }
 
   do <- function() {
     options(cli.message_class = "callr_message")
@@ -80,7 +84,9 @@ test_that("subprocess with default handler", {
 test_that("output in child process", {
   skip_on_cran()
   ## This needs callr >= 3.0.0.90001, which is not yet on CRAN
-  if (packageVersion("callr") < "3.0.0.9001") skip("Need newer callr")
+  if (packageVersion("callr") < "3.0.0.9001") {
+    skip("Need newer callr")
+  }
 
   # We need to do our own condition handling, otherwise callr will
   # handle `cli_message` and copy it to the main process.